frontend/requirements.txt
numpy==2.1.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
pypdf==5.0.0
python-dotenv==1.0.1
streamlit==1.38.0
streamlit-aggrid==1.0.5
